+++ src/bundle.ts
@@ -1,7 +1,0 @@
-import Index from './index';
-
-// only for browser
-(function() {
-  // @ts-ignore
-  window.HttpRequestMock = window.HttpRequestMock || Index
-})();